finscrape/absorbed/geo/canairy/requirements.txt
# Production requirements for Flask dashboard
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0

# Scheduling
APScheduler==3.10.4

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Web Scraping (for collectors)
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=4.9.0
pdfminer.six>=20221105

# Market Data
yfinance>=0.2.28
